Refactored how files are ordered
